Built script.js
Wrote the script for the button that runs away by following along with
the lecture recording. I added the (- btn.offsetwidth/height) to the
calculation of the new position so that the button doesn't go off the
screen. I ommitted the math.floor bit just to experiment and it seems
to work fine without it.